LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/import-assertions/invalid-type-assertion-error-expected.txt

CONSOLE MESSAGE: TypeError: Import assertion type "notARealType" is not valid
CONSOLE MESSAGE: TypeError: Import assertion type "" is not valid
CONSOLE MESSAGE: TypeError: Import assertion type "js" is not valid
CONSOLE MESSAGE: TypeError: Import assertion type "javascript" is not valid

PASS Test that invalid module type assertion leads to TypeError on window.
